pyproject.toml
[tool.poetry]
name = "pgcrow"
version = "0.7.0"
description = "Pygame framework"
authors = ["KentAugust <kentaugustone@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pygame-ce = "^2.3.1"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.5"
black = "^23.7.0"

[build-system]
requires = ["poetry-core", "setuptools"]
build-backend = "setuptools.build_meta"